Split src/preview/eviction.ts into eviction/ subdirectory (warren-d0a9)
src/preview/eviction.ts (819 lines) → src/preview/eviction/{config,types,
tick,repo,sidecar,worker}.ts; eviction.test.ts (545 lines) →
{config,tick,worker}.test.ts + test-helpers.ts. index.ts re-exports the
full public surface so call sites (server/main.ts, server/handlers/
{diagnostics,runs}.ts, diagnostics/checks.ts, preview/teardown.ts) only
update the import suffix to /index.ts.

Module layout follows the proxy/, reap/, stream/, spawn/ split
conventions (pl-9088 phases 1-7):

  - config.ts   - env constants + parsers + loadPreviewEvictionConfigFromEnv
  - types.ts    - shared types (RunPreviewRow, RunPreviewsRepo,
                  ManualTeardownClaim, tick I/O, sidecar surface)
  - tick.ts     - eviction strategy: runPreviewEvictionTick +
                  classifyRow / loadProjectOverrides / applyEviction
  - repo.ts     - drizzle-backed createRunPreviewsRepo (lifecycle
                  persistence + manual-teardown CAS)
  - sidecar.ts  - pool-backed createPoolSidecarResolver
  - worker.ts   - periodic scheduler (startPreviewEvictionWorker)

File-size budget entry dropped (each new module is well under 500
lines). The noExcessiveCognitiveComplexity grandfather entry for the
old monolith is also dropped: the four hot spots that would have
re-tripped the rule after the move are refactored via helper
extraction (tryLoadWarrenConfig+parseOverrideDuration and
stopSidecarsForRow+deleteSidecarSafely+emitEvictedEvent in tick;
lockRowForPostgres+readCurrentPreview+writeTornDown+shapeNonClaimable
in repo; buildTickInput+runTickAndCount in worker).

bun run check:all is green.
