+++ .mulch/expertise/build.jsonl
@@ -88,3 +88,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T20:03:55.922Z","evidence":{"commit":"b66f19b23a3aa212e0054d2726758dd8ffe34a2c"},"relates_to":["mx-99ad0d"],"dir_anchors":["src/server/handlers"],"content":"Handlers monolith split phase 3 (warren-599c / pl-9088 step 3): src/server/handlers/index.ts → agents.ts (203 lines), projects.ts (261), burrows.ts (116), workers.ts (89), diagnostics.ts (136), meta.ts (42); handlers/index.ts shrinks from 1194 to 420 lines and now contains only the ROUTE_TABLE composer + shared parsing helpers + API_PREFIXES/isAuthExempt/buildApiRoutes. Pattern follows phase 1 (mx-3df5c5) and phase 2 (mx-99ad0d): shared helpers re-imported via './index.ts'; ROUTE_TABLE stays in handlers/index.ts as canonical wire surface for gen:docs and gen:openapi. Route handler symbol renames (listAgents→listAgentsHandler, healthz→healthzHandler, readyz→readyzHandler, version→versionHandler) propagate through docs/http-api.md + docs/openapi.yaml — must run 'bun run gen:docs && bun run gen:openapi' after the split. Tests reorganised: handlers.test.ts → handlers/agents.test.ts + handlers/projects.test.ts + handlers/brainstorm.test.ts; handlers.preview.test.ts → handlers/index.test.ts (isAuthExempt only) + handlers/meta.test.ts (/preview/config) + handlers/runs.preview.test.ts (login + teardown + proxy preamble); handlers.burrows.test.ts and handlers.workers.test.ts move under handlers/ without splitting. agents.ts handles BOTH GET/POST /agents routes AND POST /projects/:id/agents/refresh since projectCanopyClient + refreshProjectAgents are the agents subsystem. diagnostics.ts owns /readyz with private check helpers; meta.ts holds inert /healthz + /version + /preview/config probes. Removed handlers/index.ts from biome's noExcessiveCognitiveComplexity override list; added handlers/projects.ts (runProjectTriggerHandler stays complex at 22).","id":"mx-929bbc"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T20:40:35.382Z","evidence":{"seeds":"warren-041e","commit":"a6c6a58164f2e3d58eb431548d59b1b2a5fd6bb4"},"relates_to":["mx-929bbc"],"dir_anchors":["src/runs"],"content":"Stream bridge split (warren-041e / pl-9088 step 5): src/runs/stream.ts (911 lines) → src/runs/stream/ subdirectory. Modules: bridge.ts (bridgeRunStream pump, retains noExcessiveCognitiveComplexity override at score 63), terminal-detect.ts (detectRuntimeTerminal + isPiAgentEnd), run-state-poller.ts (warren-6596 fallback), stats.ts (pi+claude cost persistence: snapshotStats/persistPiStatsDelta/persistInStreamUsage), recover.ts (recoverActiveRunStreams + ActiveBridge types), types.ts (shared interfaces incl. SessionStats so usage-aggregate can re-export without runtime cycle). Tests split into bridge.test.ts (event flow + terminal detection + b1a9 ghost-run), run-state-poller.test.ts (warren-6596), stats.test.ts (PiStatsClient + in-stream pi + in-stream claude), recover.test.ts, plus terminal-detect.test.ts (moved from src/runs/stream.detect.test.ts), backed by test-helpers.ts (makePool, evt, source, piAgentEnd, piTurnEnd, claudeResult, seedBridgeRun). Callers updated to import from ./stream/index.ts (manage.ts, cancel.ts, runs/index.ts, usage-aggregate.ts, reap/types.ts). file-size budgets for stream.{ts,test.ts} removed; biome noExcessiveLinesPerFunction override for stream.test.ts removed. All files now ≤ 500 lines (largest: stats.test.ts at 497).","id":"mx-c7ca2b"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T21:37:07.839Z","evidence":{"commit":"b66f19b2","seeds":"warren-a2b4"},"relates_to":["mx-3df5c5","mx-929bbc"],"dir_anchors":["src/server/handlers"],"content":"Handlers monolith split phase 2 (warren-a2b4 / pl-9088 step 2) — merged in commit b66f19b2. Extracted src/server/handlers/runs.ts (789 lines: listRunsHandler, getRunHandler, createRunHandler, postRunMessageHandler, steerRunHandler, cancelRunHandler, previewLoginHandler, previewTeardownHandler, streamRunEventsHandler, listCostAnalyticsHandler + streaming helpers bridgeAbort/asNdjsonStream/eventToNdjson) and src/server/handlers/plan-runs.ts (484 lines: createPlanRunHandler, listPlanRunsHandler, getPlanRunHandler, cancelPlanRunHandler, streamPlanRunEventsHandler + tailPlanRunEvents). Tests: src/server/handlers/runs.test.ts (988 lines) extracts POST /runs spawn/plot_id/interactive + /runs/:id/messages + GET /runs/:id/events describes from handlers.test.ts; src/server/handlers/plan-runs.test.ts (1318 lines) relocated from src/server/handlers.plan-runs.test.ts, biome noExcessiveLinesPerFunction override removed by splitting the monolithic POST /plan-runs describe into three siblings sharing setupPlanRunFixture(). Reconfirms playbook from phase 1 (mx-3df5c5) and phase 3 (mx-929bbc): ROUTE_TABLE stays in handlers/index.ts as canonical wire surface; shared parsing helpers exported from handlers/index.ts.","id":"mx-89e2c7"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-27T22:16:53.909Z","evidence":{"seeds":"warren-d0a9","commit":"b0c91170cc71f3a7fea42c3d0eef4bdb2912e145"},"tags":["refactor","file-size-ratchet","pl-9088"],"relates_to":["mx-c7ca2b"],"dir_anchors":["src/preview"],"content":"Preview eviction split (warren-d0a9 / pl-9088 step 8): src/preview/eviction.ts (819 lines) → src/preview/eviction/{config,types,tick,repo,sidecar,worker}.ts; eviction.test.ts (545) → {config,tick,worker}.test.ts + test-helpers.ts. index.ts re-exports the public surface so call sites (server/main.ts, handlers/{diagnostics,runs}.ts, diagnostics/checks.ts, preview/teardown.ts) only change the suffix to /index.ts. Dropped the file-size budget entry and the noExcessiveCognitiveComplexity grandfather entry for the old monolith; new complexity hot spots refactored via helper extraction (tryLoadWarrenConfig+parseOverrideDuration + stopSidecarsForRow+deleteSidecarSafely+emitEvictedEvent in tick; lockRowForPostgres+readCurrentPreview+writeTornDown+shapeNonClaimable in repo; buildTickInput+runTickAndCount in worker).","id":"mx-5c4f4d"}
