+++ src/preview/eviction/tick.test.ts
@@ -1,113 +1,13 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { ValidationError } from "../core/errors.ts";
-import { type AnyWarrenDb, openDatabase } from "../db/client.ts";
-import { createRepos, type Repos } from "../db/repos/index.ts";
-import { isPostgresTestEnabled, withDb } from "../db/testing.ts";
-import { RunEventBroker } from "../runs/events.ts";
-import { createWarrenConfigCache, type LoadedWarrenConfig } from "../warren-config/index.ts";
-import {
-	createRunPreviewsRepo,
-	DEFAULT_IDLE_TTL_MS,
-	DEFAULT_MAX_LIFETIME_MS,
-	DEFAULT_MAX_LIVE,
-	type EvictionReason,
-	loadPreviewEvictionConfigFromEnv,
-	type PreviewEvictionConfig,
-	type RunPreviewsRepo,
-	runPreviewEvictionTick,
-	type SidecarClient,
-	type SidecarResolver,
-	startPreviewEvictionWorker,
-	WARREN_PREVIEW_EVICTION_DISABLED_ENV,
-	WARREN_PREVIEW_EVICTION_TICK_MS_ENV,
-	WARREN_PREVIEW_IDLE_TTL_ENV,
-	WARREN_PREVIEW_MAX_LIFETIME_ENV,
-	WARREN_PREVIEW_MAX_LIVE_ENV,
-} from "./eviction.ts";
-
-describe("loadPreviewEvictionConfigFromEnv", () => {
-	test("defaults match SPEC §11.L", () => {
-		expect(loadPreviewEvictionConfigFromEnv({})).toEqual({
-			idleTtlMs: DEFAULT_IDLE_TTL_MS,
-			maxLifetimeMs: DEFAULT_MAX_LIFETIME_MS,
-			maxLive: DEFAULT_MAX_LIVE,
-			tickMs: 10_000,
-			disabled: false,
-		});
-	});
-
-	test("parses idle TTL + max-lifetime durations", () => {
-		const cfg = loadPreviewEvictionConfigFromEnv({
-			[WARREN_PREVIEW_IDLE_TTL_ENV]: "5m",
-			[WARREN_PREVIEW_MAX_LIFETIME_ENV]: "2h",
-			[WARREN_PREVIEW_MAX_LIVE_ENV]: "5",
-			[WARREN_PREVIEW_EVICTION_TICK_MS_ENV]: "1500",
-		});
-		expect(cfg.idleTtlMs).toBe(5 * 60_000);
-		expect(cfg.maxLifetimeMs).toBe(2 * 3_600_000);
-		expect(cfg.maxLive).toBe(5);
-		expect(cfg.tickMs).toBe(1500);
-	});
-
-	test("WARREN_PREVIEW_EVICTION_DISABLED toggles", () => {
-		expect(
-			loadPreviewEvictionConfigFromEnv({ [WARREN_PREVIEW_EVICTION_DISABLED_ENV]: "1" }).disabled,
-		).toBe(true);
-		expect(
-			loadPreviewEvictionConfigFromEnv({ [WARREN_PREVIEW_EVICTION_DISABLED_ENV]: "true" }).disabled,
-		).toBe(true);
-		expect(
-			loadPreviewEvictionConfigFromEnv({ [WARREN_PREVIEW_EVICTION_DISABLED_ENV]: "0" }).disabled,
-		).toBe(false);
-	});
-
-	test("malformed env values fail loudly (incl. junk-suffix warren-da37)", () => {
-		const cases: Array<[string, string]> = [
-			[WARREN_PREVIEW_IDLE_TTL_ENV, "garbage"],
-			[WARREN_PREVIEW_MAX_LIVE_ENV, "0"],
-			[WARREN_PREVIEW_MAX_LIVE_ENV, "-1"],
-			[WARREN_PREVIEW_MAX_LIVE_ENV, "5abc"], // warren-da37: junk suffix
-			[WARREN_PREVIEW_EVICTION_TICK_MS_ENV, "1500x"], // warren-da37
-		];
-		for (const [n, v] of cases)
-			expect(() => loadPreviewEvictionConfigFromEnv({ [n]: v })).toThrow(ValidationError);
-	});
-});
-
-interface FakeSidecars {
-	resolver: SidecarResolver;
-	calls: Array<{ burrowId: string; sidecarId: string }>;
-	listings: Map<string, string[]>;
-}
-
-function fakeSidecars(): FakeSidecars {
-	const listings = new Map<string, string[]>();
-	const calls: Array<{ burrowId: string; sidecarId: string }> = [];
-	const resolver: SidecarResolver = async (_burrowId): Promise<SidecarClient> => ({
-		list: async (id) => (listings.get(id) ?? []).map((sid) => ({ id: sid })),
-		delete: async (id, scid) => {
-			calls.push({ burrowId: id, sidecarId: scid });
-			const cur = listings.get(id) ?? [];
-			listings.set(
-				id,
-				cur.filter((s) => s !== scid),
-			);
-		},
-	});
-	const seededResolver: SidecarResolver = async (burrowId) => {
-		if (!listings.has(burrowId)) listings.set(burrowId, [`sc_${burrowId}`]);
-		return resolver(burrowId);
-	};
-	return { resolver: seededResolver, calls, listings };
-}
-
-const BASE_CONFIG: PreviewEvictionConfig = {
-	idleTtlMs: 30 * 60_000,
-	maxLifetimeMs: 8 * 3_600_000,
-	maxLive: 20,
-	tickMs: 10_000,
-	disabled: false,
-};
+import type { AnyWarrenDb } from "../../db/client.ts";
+import { createRepos, type Repos } from "../../db/repos/index.ts";
+import { isPostgresTestEnabled, withDb } from "../../db/testing.ts";
+import { RunEventBroker } from "../../runs/events.ts";
+import { createWarrenConfigCache, type LoadedWarrenConfig } from "../../warren-config/index.ts";
+import { createRunPreviewsRepo } from "./repo.ts";
+import { BASE_CONFIG, fakeSidecars } from "./test-helpers.ts";
+import { runPreviewEvictionTick } from "./tick.ts";
+import type { EvictionReason, RunPreviewsRepo } from "./types.ts";
 
 function tickSuite(dialect: "sqlite" | "postgres"): void {
 	describe(`runPreviewEvictionTick (${dialect})`, () => {
@@ -471,75 +371,3 @@ tickSuite("sqlite");
 if (isPostgresTestEnabled()) {
 	tickSuite("postgres");
 }
-
-describe("startPreviewEvictionWorker", () => {
-	let db: AnyWarrenDb;
-	let repos: Repos;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-		await repos.agents.upsert({ name: "agent", renderedJson: { sections: {} } });
-		await repos.projects.create({
-			gitUrl: "https://github.com/x/y.git",
-			localPath: "/data/projects/x/y",
-			defaultBranch: "main",
-		});
-	});
-
-	afterEach(async () => {
-		await db.close();
-	});
-
-	test("runOnce fires the tick and increments tick count", async () => {
-		const sidecars = fakeSidecars();
-		const configs = createWarrenConfigCache({
-			load: async () => ({
-				triggers: null,
-				defaults: null,
-				prTemplate: null,
-				errors: [],
-				warnings: [],
-			}),
-		});
-		const handle = startPreviewEvictionWorker({
-			db,
-			repos,
-			burrowClientPool: undefined as never,
-			warrenConfigs: configs,
-			config: { ...BASE_CONFIG, disabled: true },
-			resolveSidecar: sidecars.resolver,
-		});
-
-		const result = await handle.runOnce();
-		expect(result).not.toBeNull();
-		expect(handle.tickCount()).toBe(1);
-		await handle.stop();
-	});
-
-	test("disabled mode skips setInterval", async () => {
-		let timersStarted = 0;
-		const handle = startPreviewEvictionWorker({
-			db,
-			repos,
-			burrowClientPool: undefined as never,
-			warrenConfigs: createWarrenConfigCache({
-				load: async () => ({
-					triggers: null,
-					defaults: null,
-					prTemplate: null,
-					errors: [],
-					warnings: [],
-				}),
-			}),
-			config: { ...BASE_CONFIG, disabled: true },
-			setInterval: () => {
-				timersStarted += 1;
-				return {};
-			},
-			clearInterval: () => {},
-		});
-		expect(timersStarted).toBe(0);
-		await handle.stop();
-	});
-});
